# static/swagger.yaml
swagger: '2.0'
info:
  description: >
    DEBiasing embeddings implicitly and explicitly - This is a tool providing
    bias evaluation scores and metrices and debiasing models for implicit and
    explicit bias specifications on embedding spaces. 

    http://debie.informatik.uni-mannheim.de:8000/REST
  version: 0.2.0
  title: DEBIE
  contact:
    email: nfriedri@mail.uni-mannheim.de
  license:
    name: MIT License
    url: 'https://github.com/nfriedri/debie-backend/blob/master/LICENSE'
...
paths:
  /vectors/single:
    get:
      tags:
        - Vector Retrieval
      summary: Retrieves the vector representation of a word
      description: |
        The system returns the vector representation of a input word
        out of a specified embedding space.
      produces:
        - application/json
      parameters:
        - in: query
          name: word
          description: input word for vector retrieval
          required: true
          type: string
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
      responses:
        '200':
          description: Succesfully found vector representation
          schema:
            type: array
            items:
              $ref: '#/definitions/SingleVectorResult'
        '400':
          description: bad input parameter
        '404':
          description: NOT FOUND - input value not found in database
  /vectors/multiple:
    post:
      tags:
        - Vector Retrieval
      summary: Retrieves the vector representations of a list of words
      description: >
        The system returns the vector representations of a list of words
        separated by blank spaces out of a specified embedding space.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: List of words separated by blank spaces
          required: true
          schema:
            $ref: '#/definitions/WordList'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file.
          required: false
          type: string
        - in: query
          name: lower
          description: lowercases all letters of the given input word
          required: false
          default: false
          type: string
      responses:
        '200':
          description: Succesfully found vector representation
          schema:
            type: array
            items:
              $ref: '#/definitions/MultipleVectorResult'
        '400':
          description: bad input parameter
  /augmentations/single:
    get:
      tags:
        - Augmentation Retrieval
      summary: Retrieves augmentations for a given input word
      description: >
        The system returns a list of 4 augmentations for the input word out of a
        postspecialized embedding space.
      produces:
        - application/json
      parameters:
        - in: query
          name: word
          description: input word for vector retrieval
          required: true
          type: string
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
      responses:
        '200':
          description: Succesfully found augmentations
          schema:
            type: array
            items:
              $ref: '#/definitions/SingleAugmentationResult'
        '400':
          description: failure
  /augmentations/multiple:
    post:
      tags:
        - Augmentation Retrieval
      summary: Retrieves the augmentations of a list of words
      description: >
        The system returns the computed postspecialized augmentations of a list
        of words separated by blank spaces.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: List of words separated by blank spaces
          required: true
          schema:
            $ref: '#/definitions/WordList'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
      responses:
        '200':
          description: Succesfully found vector representation
          schema:
            type: array
            items:
              $ref: '#/definitions/MultipleAugmentationResult'
        '400':
          description: bad input parameter
  /bias-evaluation/all:
    post:
      tags:
        - Bias Evaluation
      summary: Computes the main bias evaluation scores
      description: >
        computes bias evaluation scores ECT, BAT, WEAT, K-Means++ and SVM on the
        given bias specification.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: >-
            contains the explicit bias specification, consisting out of the two
            target term sets T1 and T2 and the two attribute term sets A1 and
            A2.
          required: true
          schema:
            $ref: '#/definitions/BiasSpecification'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file. If no file is
            uploaded, either set the param to false or leave it out.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
        - in: query
          name: json
          description: >-
            set this parameter to true if the vector data should be retrieved
            out of the request's body. Else set it to false or leave it out.
          required: false
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/BiasEvaluationALL'
        '400':
          description: failure
  /bias-evaluation/ect:
    post:
      tags:
        - Bias Evaluation
      summary: Computes the ECT-score
      description: >
        computes the ECT-score (Embedding Coherence Test) with p-value on the
        given bias specification.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: >-
            contains the explicit bias specification, consisting out of the two
            target term sets T1 and T2 and the two attribute term sets A1 and
            A2.
          required: true
          schema:
            $ref: '#/definitions/BiasSpecification'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file. If no file is
            uploaded, either set the param to false or leave it out.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
        - in: query
          name: json
          description: >-
            set this parameter to true if the vector data should be retrieved
            out of the request's body. Else set it to false or leave it out.
          required: false
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/BiasEvaluationECT'
        '400':
          description: failure
  /bias-evaluation/bat:
    post:
      tags:
        - Bias Evaluation
      summary: Computes the BAT-score
      description: >
        computes the BAT-score (Bias Analogy Test) on the given bias
        specification.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: >-
            contains the explicit bias specification, consisting out of the two
            target term sets T1 and T2 and the two attribute term sets A1 and
            A2.
          required: true
          schema:
            $ref: '#/definitions/BiasSpecification'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file. If no file is
            uploaded, either set the param to false or leave it out.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
        - in: query
          name: json
          description: >-
            set this parameter to true if the vector data should be retrieved
            out of the request's body. Else set it to false or leave it out.
          required: false
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/BiasEvaluationBAT'
        '400':
          description: failure
  /bias-evaluation/weat:
    post:
      tags:
        - Bias Evaluation
      summary: Computes the WEAT-score
      description: >
        computes the WEAT (Word Embedding Association Test) effect size with
        p-value on the given bias specification.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: >-
            contains the explicit bias specification, consisting out of the two
            target term sets T1 and T2 and the two attribute term sets A1 and
            A2.
          required: true
          schema:
            $ref: '#/definitions/BiasSpecification'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file. If no file is
            uploaded, either set the param to false or leave it out.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
        - in: query
          name: json
          description: >-
            set this parameter to true if the vector data should be retrieved
            out of the request's body. Else set it to false or leave it out.
          required: false
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/BiasEvaluationWEAT'
        '400':
          description: failure
  /bias-evaluation/kmeans:
    post:
      tags:
        - Bias Evaluation
      summary: Computes the K-Means++ score
      description: |
        computes the K-Means++ score on the given bias specification.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: >-
            contains the explicit bias specification, consisting out of the two
            target term sets T1 and T2 and the two attribute term sets A1 and
            A2.
          required: true
          schema:
            $ref: '#/definitions/BiasSpecification'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file. If no file is
            uploaded, either set the param to false or leave it out.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
        - in: query
          name: json
          description: >-
            set this parameter to true if the vector data should be retrieved
            out of the request's body. Else set it to false or leave it out.
          required: false
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/BiasEvaluationKMEANS'
        '400':
          description: failure
  /bias-evaluation/svm:
    post:
      tags:
        - Bias Evaluation
      summary: Computes the SVM-score
      description: >
        computes the SVM-score (Support Vector Machine) on the given bias
        specification.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: >-
            contains the explicit bias specification, consisting out of the two
            target term sets T1 and T2 and the two attribute term sets A1 and
            A2.
          required: true
          schema:
            $ref: '#/definitions/BiasSpecification'
        - in: query
          name: space
          description: >-
            embedding space out of which the vector representation will be
            retrieved
          required: true
          default: fasttext
          type: string
        - in: query
          name: uploaded
          description: >-
            contains the filename of the uploaded embedding space in case the
            data should be retrieved from an uploaded file. If no file is
            uploaded, either set the param to false or leave it out.
          required: false
          type: string
        - in: query
          name: lower
          description: Lowercases all letters of the given input word
          required: false
          default: false
          type: string
        - in: query
          name: json
          description: >-
            set this parameter to true if the vector data should be retrieved
            out of the request's body. Else set it to false or leave it out.
          required: false
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/BiasEvaluationSVM'
        '400':
          description: failure
...
definitions:
  WordList:
    type: object
    required:
      - Words
    properties:
      Words:
        type: string
        format: utf-8
        example: car truck plane ship boat
...
  SingleVectorResult:
    type: object
    required:
      - Vector
      - NotFound
    properties:
      Vector:
        type: array
        items:
          type: object
          properties:
            word1:
              type: number
              example: 0.12345 0.4568 0.123456 0.122 ...
      NotFound:
        type: string
        example: '[]'
  MultipleVectorResult:
    type: object
    required:
      - Vector
      - NotFound
    properties:
      Vector:
        type: array
        items:
          type: object
          properties:
            word1:
              type: number
              example: 0.12345 0.4568 0.123456 0.122 ...
            word2:
              type: number
              example: 0.12345 0.4568 0.123456 0.122 ...
            word3:
              type: number
              example: 0.12345 0.4568 0.123456 0.122 ...
      NotFound:
        type: string
        example: '[''word4'', ''word5'']'
  SingleAugmentationResult:
    type: object
    required:
      - Augmentations
      - NotPostspecialized
    properties:
      Augmentations:
        type: object
        properties:
          word1:
            type: array
            items:
              type: string
              example: word2 word3 word4
      NotPostspecialized:
        type: string
        example: '[]'
...
host: 'wifo5-29.informatik.uni-mannheim.de:8000'
basePath: /REST/
schemes:
  - http